% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTalk}
\alias{getTalk}
\title{Download a Talk}
\usage{
getTalk(url)
}
\arguments{
\item{url}{URL of the talk to be downloaded}
}
\value{
A character string giving the full text of the talk.  Note that
references are not included.
}
\description{
Download a talk from lds.org and process it into plain text.
}
\details{
The complete HTML code from the page is downloaded.  A series of
regular expressions is used to parse the HTML into a plain text version of
the talk.  Using regular expressions is not generally recommended for
parsing HTML. Unfortunately, the HTML code doesn't appear to play well with
more advanced forms of parsing.  The regular expressions seem to work well
here mostly because talk paragraphs start with the tag
\code{<p uri= ... id="pid[num]">}.  HTML rendered characters are converted
to plain text and tags such as links, italic, and bold are removed.
}
\examples{
# Download the first talk of the 1971 General Conference
getTalk('https://www.lds.org/general-conference/1971/04/out-of-the-darkness?lang=eng')
}
\author{
Benjamin Nutter
}

